New port: textproc/py-pikepdf
pikepdf is a Python library for reading and writing PDF files. It's based on
QPDF, a powerful PDF manipulation and repair library.

pikepdf is similar to PyPDF2 and pdfrw - it provides low level access to PDF
features and allows editing and content transformation of existing PDFs.

Some knowledge of the PDF specification may be helpful. It does not have the
capability to render a PDF to image.

WWW: https://github.com/pikepdf/pikepdf


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@506412 35697150-7ecd-e111-bb59-0022644237b5