graphics/drm-fbsd12.0-kmod: Mark only for 12
Change the version check in graphics/drm-fbsd12.0-kmod to make it only
compile on FreeBSD 12, not 13 and later, since it fails to build on 12.0.
graphics/drm-current-kmod is the same driver versions, but with some
differences due to differences in 12 and current.

MFH:		2019Q3 (implicit, drm-drivers blanket)
Sponsored by:	B3 Init


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@506712 35697150-7ecd-e111-bb59-0022644237b5