Add p5-URI-Fast 0.49
URI::Fast is a faster alternative to URI. It is written in C and provides basic
parsing and modification of a URI.

URI is an excellent module; it is battle-tested, robust, and handles many edge
cases. As a result, it is rather slower than it would otherwise be for more
trivial cases, such as inspecting the path or updating a single query parameter.

WWW: https://metacpan.org/release/URI-Fast