graphics/kanshi: add new port
kanshi allows you to define output profiles that are automatically
enabled and disabled on hotplug. For instance, this can be used to
turn a laptop's internal screen off when docked.

This is a Wayland equivalent for tools like autorandr. kanshi can be
used on Wayland compositors supporting wlr-output-management protocol.

Sway users: kanshi requires Sway 1.2 or later.

https://wayland.emersion.fr/kanshi/


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@508585 35697150-7ecd-e111-bb59-0022644237b5