Update to 3.8.0.0-rc2
# Changelog
All notable changes to GNU Radio will be documented in this file.

The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html), starting with version 3.7.12.0.

Older Logs can be found in `docs/RELEASE-NOTES-*`.

## [3.8.0.0-rc2] - 2019-07-20

This is the second release candidate for GNU Radio 3.8.

### Fixed
#### Project Scope
- Last minute patch lead to SWIG libraries not getting installed into the right Python paths
#### gr-utils
- `gr_modtool`: Paths with spaces in them used to break

### Changed
#### GRC
- Copy & Paste behaviour now like on 3.7: appending numbers to block IDs only on conflicts

## [3.8.0.0-rc1] - 2019-07-17

This is the release candidate for GNU Radio 3.8.

For brevity reasons, this document does NOT contain a complete CHANGELOG
comparing 3.8.0.0 to 3.7.0, because the development model changed underway, and
there's no human-feasible way to detangle the changes from more than six years
of development.

The major user-relevant changes will be noted here in the actual release.

What we can do here is point out that we've got a glorious amount of 215
individual contributors that added at least one commit since 3.7.0.

This is a **big** step towards the future of the world's most popular FOSS SDR
framework.

Let us ride eternal, shiny and chrome.

Pet Portlint

PR:		ports/239345
Submitted by:	mike.english@gmail.com