Update devel/cmake to 3.15.2 (latest upstream release)
Release notes for the version updates are at:
	https://blog.kitware.com/cmake-3-15-0-available-for-download/
	https://blog.kitware.com/cmake-3-15-1-available-for-download/
	https://blog.kitware.com/cmake-3-15-2-available-for-download/

For FreeBSD, not much interesting has changed. Swift language support
might be useful.

A few ports were fixed as part of this exp-run (libemos not really
related to the CMake update, but in general). Thanks to antoine
for the exp-runs, tcberner for doing most of the prep-work.

PR:		239273
Submitted by:	tcberner