New port: ftp/curlie: Frontend to cURL that adds the ease of use of HTTPie
If you like the interface of HTTPie but miss the features of curl, curlie
is what you are searching for. Curlie is a frontend to curl that adds the
ease of use of httpie, without compromising on features and performance.
All curl options are exposed with syntax sugar and output formatting
inspired from httpie.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@509081 35697150-7ecd-e111-bb59-0022644237b5