Add py-django-hierarkey 1.0.3
django-hierarkey -- Hierarchical key-value store

This package allows you to attach a key-value store to a model, e.g. to store
preferences of an user or a customer. The package supports arbitrary datatypes,
defaults and model hierarchies, i.e. you can define a different model instance
as your instance's parent and the values of the parent instance will be used as
default values for the child instances.

This approach has been in use in pretix for quite a while, so it has been tested
in production.

This project is maintained by Raphael Michel <mail@raphaelmichel.de>. See the
AUTHORS file for a list of all the awesome folks who contributed to this
project.

WWW: https://github.com/raphaelm/django-hierarkey

- While I'm here, add NO_ARCH

PR:		239059
Submitted by:	Alexander Sieg <alex@xanderio.de>