net/libdmapsharing: Explicitly disable the dependency on devel/check
By default, the configure script of net/libdmapsharing detects
devel/check automatically.  If devel/check is installed,
net/libdmapsharing will link to it.  It can cause problems because
devel/check is not declared as a dependency of net/libdmapsharing,
so libdmapsharing-3.0.so will be broken if devel/check is removed.

$ ldd /usr/local/lib/libdmapsharing-3.0.so
...
libcheck.so.0 => not found (0)
...

Since devel/check is just a unit test framework, it should be safe
to disable it unconditionally.

PR:		240195
Submitted by:	Ting-Wei Lan <lantw44@gmail.com>