utouch - driver for absolute USB HID mouses emulated by some virtualization systems like Virtual Box, BHyve, e.t.c. It uses evdev protocol to communicate with userland applications like libinput and xf86-input-evdev.
The driver should be installed in to the guest FreeBSD system. Host system
should be configured to emulate mouse as a single-touch USB tablet.

WWW: https://github.com/wulf7/utouch

It is a temporary solution until absolute coords support is added to ums(4)
driver. That is why it is not imported into the base system.

PR:		240261
Approved by:	koobs