textproc/docbookrx: create port
DocBookRx is the prescription you need to get rid of your DocBook pain. This
tool converts DocBook XML to AsciiDoc.

DocBookRx is the start of a DocBook to AsciiDoc converter written in Ruby. This
converter is far from perfect at the moment and some of the conversion is done
hastily. The plan is to evolve it into a robust library for performing this
conversion in a reliable way.

WWW: https://github.com/asciidoctor/docbookrx