Upgrade editors/openoffice-devel to upstream git hash d871312c80.
Since upstream has migrated from svn to git, use seconds from epoch of
the last commit as the minor component of PORTVERSION.

The build wants to include the git hash value in the build for display
on the Help -> About popup, but it can't determine that when building
from a source tarball, as opposed to a git checkout.  Patch around that
issue until upstream implements a way to include this info into the
tarball and pick it up from there.

Add xorg to USES for correctness.