Add net/dual-dhclient-daemon port.
This serves the same purpose as net/dual-dhclient, but does it
better: While net/dual-dhclient is a trivial shell script which
launches two dhclients, this is a daemon which
(a) can be signalled by /etc/rc.d/dhclient stop, and
(b) relays that signal to the two "child" dhclient daemons.

This is deliberately not replacing the net/dual-dhclient port
since in the unlikely event that something is broken I don't
want to see pkg automatically upgrading people; but EC2 AMIs
will be making use of this new port and I intend to remove the
old dual-dhclient port at some point.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@513630 35697150-7ecd-e111-bb59-0022644237b5