Limit USES+=compiler:c11 to powerpc* and FreeBSD < 13.
MIPS regressed in r513614 because it does not have a C11
compiler. Fetchmail 6.4.1 is C89 code, one patch we provide
makes it C99 (it requires the long long int type).

The base compiler on powerpc 11/12 does not yield an executable,
but MIPS on the other hand does not offer a C11 compiler,
so for now, revert anything but powerpc before FreeBSD 13
to use its base compiler.  FreeBSD 13 is exempt and also uses
the base compiler, because the ARCH guys intend to merge LLVM 9
soon.

PR:		241031
Reviewed by:	pkubaj@
Approved by:	blanket (fix recent regression from r513614)


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@513664 35697150-7ecd-e111-bb59-0022644237b5