Try to fix runtime graphics/digikam.
With this patch applied to git master, digikam starts and seems
to work. It still crashes on exit, though. That's an improvement
on crashes-before-startup.

The problem is described in the patch and in the PR: digikam
bundles all kinds of stuff (which packagers have been complaining
about for years) which breaks -- in this case, bundling internals
of Exiv2.

If this works (leaving the PR open) it will need an MFH.

PR:		240466