Drop LLD_UNSAFE=yes.
The port is building fine on 12.0 and HEAD without setting this. The port's
build system is still wrong (i.e. the port still builds the same moc-generate
code in duplicate files), but lld seems to handle this case well.

A proper fix was sent upstream in https://phabricator.kde.org/D24609.

I am not bumping PORTREVISION because this only changes build dependencies.

PR:		228640


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@514392 35697150-7ecd-e111-bb59-0022644237b5