Divorce the ARM bare-metal GCC ports from the FreeBSD GCC xtoolchain ports.
In the past, changes to the FreeBSD GCC xtoolchain ports have broken
the ARM bare-metal GCC ports by accident.  In addition, we'd like to
move to having multiple FreeBSD GCC xtoolchain ports for different GCC
versions.  It seems best if we make the two groups of ports
independent so that changes to one do not affect the other.

Reviewed by:	manu
Differential Revision:	https://reviews.freebsd.org/D22064


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@515276 35697150-7ecd-e111-bb59-0022644237b5