security/cyrus-sasl2: unbreak building with stock OpenSSL for stable/11
Sendmail bundled with FreeBSD has SASL support and the Handbook tells
how to rebuild the Sendmail with SASL enabled if you have installed
cyrus-sasl2 that links with OpenSSL's libcrypto.

Sendmail uses old OpenSSL 1.0.2 API, so cyrus-sasl2 should be built
with stock libcrypto even if newer OpenSSL is installed for Ports.

This change adds new option SSL to the port (enabled by default).
If disabled, cyrus-sasl2 is built witch stock libcrypto not depending
on ports version of OpenSSL.

PORTREVISION not changed as default build is not affected.