qemu-user-static: fix the RFSPAWN fix + kevent(2) emulation
The RFSPAWN fix included a bogus dependency on qemu-user-static being built
against headers that included the RFSPAWN flag. This is incorrect in
emulation-land, where we must assume the target may not match the host.

kevent(2) emulation was also not preserving event udata coming from the
kernel; the fix for this is als included.

Approved by:	antoine (ports), sbruno (maintainer, implicit)
MFH:		2019Q4