x11-wm/chamfer: add new port
Chamferwm is a dynamic tiling window manager with a built-in Vulkan
based compositor. Besides aiming to be a fully featured and highly
configurable window manager, significant effort is put to provide a
complete control over the appearance of the desktop. With its shader
based customization system, Chamferwm enables rendering of arbitrary
window decorations, borders and effects, maximizing the potential for
visual personalization.

Window manager:
- Dynamic horizontal and vertical tiling with gaps and stacking
- Specify container size restrictions, overlap when necessary
- Resize and translate individual containers in their place while
  keeping the surrounding layout
- Floating containers and automatic dialog, dock, widget etc. handling
- Yank and paste containers to move them within the tree hierarchy
  along with typical move operations
- Configuration with python, scriptable behaviour with client and
  container specific callbacks
- Fully keyboard controllable

Compositor:
- Vulkan renderer
- Arbitrary window decorations and borders with user supplied shaders
- Per-client materials
- Optional, alternatively use any other external compositor

https://jaelpark.github.io/chamferwm-docs/