Add sysutils/duplicity-devel: Backup tool that uses librsync and GnuPG
This is the current development release.

Duplicity backs directories by producing encrypted tar-format volumes and
uploading them to a remote or local file server. Because duplicity uses
librsync, the incremental archives are space efficient and only record the
parts of files that have changed since the last backup. Because duplicity
uses GnuPG to encrypt and/or sign these archives, they will be safe from
spying and/or modification by the server.

WWW: http://duplicity.nongnu.org


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@517676 35697150-7ecd-e111-bb59-0022644237b5