Update from 5.0.6 to 5.0.7.
<ChangeLog>

Upgrade urgency HIGH: many issues fixed, some may have an impact.

Hi all, Redis 5.0.7 fixes a number of bugs, none is very critical, however
there are a few that may have an impact. It's a good idea to upgrade.
There are fixes in the area of replication from modules commands and
callbacks, AOF fsync (non critical issue), memory leaks (very rare and small),
streams beahvior (non critical), and a potential crash in commands
processing multiple keys at the same time that is there for years, and happens
very rarely, but is not impossible to trigger.

</ChangeLog>


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@518108 35697150-7ecd-e111-bb59-0022644237b5