x11/py-i3ipc: add new port
i3's interprocess communication (or ipc) is the interface i3wm uses to
receive commands from client applications such as i3-msg. It also
features a publish/subscribe mechanism for notifying interested
parties of window manager events.

i3ipc-python is a Python library for controlling the window
manager. This project is intended to be useful for general scripting,
and for applications that interact with the window manager like status
line generators, notification daemons, and window pagers.

http://i3ipc-python.readthedocs.io/