- New port: www/thirtybees
thirty bees is a matured e-commerce solution which once started as a fork of
PrestaShop 1.6.1.11 and is still compatible with (almost) all PS 1.6 modules.
Its focus is on stability, correctness and reliability of the rich feature set,
to allow merchants to focus on growing their business.

WWW: https://thirtybees.com