devel/racer: Assign to rust@
Racer often needs to be updated when we update lang/rust since it's
tied to nightly features.  We have always just updated this port
in the past and will do so again when updating to Rust 1.40.0.
Let's make it official and assign the port to the Rust team.

PR:		242685
Approved by:	beyert@cs.ucr.edu (maintainer)


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@520360 35697150-7ecd-e111-bb59-0022644237b5