Add a new port for an external FreeBSD toolchain based on GCC 9.
This port provides external GCC toolchains for GCC 9.2.0 for the aarch64,
amd64, i386, mips, mips64, powerpc, powerpc64, riscv64, and sparc64
platforms.

Note that at present I am not yet able to build a full FreeBSD system
on any of these platforms, but it will be easier to debug the remaining
issues (which may be issues to fix in FreeBSD rather than the port) with
the packages available to other developers.

Reviewed by:	bapt, cem
Approved by:	portmgr (bapt)
Differential Revision:	https://reviews.freebsd.org/D22863