tmux-mem-cpu-load: Fix memory calculation, and remove tmux dependency
This adds a patch from fluffy that fixes the memory usage calculation, which
was previously wildly incorrect (on my box: before 484, after 16889). See
https://github.com/thewtex/tmux-mem-cpu-load/pull/61

Also, I'm removing the run-time dependency on tmux. Although this program
is designed to be used in the tmux status bar, there's no reason it can't
be used in any other status bar, and it doesn't actually use tmux in any
way.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@520651 35697150-7ecd-e111-bb59-0022644237b5