Add py-cachecontrol 0.12.6
CacheControl is a port of the caching algorithms in httplib2 for use with
requests session object.

It was written because httplib2's better support for caching is often mitigated
by its lack of thread safety. The same is true of requests in terms of caching.

WWW: https://github.com/ionrock/cachecontrol