Update audio/rosegarden to latest upstream commit
Release notes are at
	https://www.rosegardenmusic.com/wiki/dev:19.12

There's two handfuls of minor bugfixes and some new auto-scrolling
features.

I notice now that Rosegarden crashes (this version, also the version
in ports, 18.12) if you double-click on a fragment, so it looks like
there are very few heavy users of the application (or I'm doing it
wrong).

PR:		242957