Update devel/cmake to latest upstream release, 3.16.2
This new release has support for the latest Boost, Objective-C
and various small improvements. The release notes are at:
	https://cmake.org/cmake/help/v3.16/release/3.16.html

Two ports could not be persuaded to build with this new release
(math/openmesh science/fleur) and are marked BROKEN. Other build
fixes have already been done in-tree.

Thanks antoine@ for the exp-runs (multiple).
Thanks tcberner@ for prep-work (most of the work, actually).

PR:		242254