New port: devel/msitools
msitools is a set of programs to inspect and build Windows Installer (.MSI)
files. It is based on libmsi, a portable library to read and write .MSI files.
libmsi in turn is a port of (and a subset of) Wine's implementation of the
Windows Installer.

msitools plans to be a solution for packaging and deployment of cross-compiled
Windows applications.

WWW: https://wiki.gnome.org/msitools


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@522355 35697150-7ecd-e111-bb59-0022644237b5