(Readdition of net-p2p/btpd which was removed on 514600)
BTPD is a bittorrent client consisting of a daemon and client commands,
which can be used to read and/or manipulate the daemon state. The daemon
is capable of running several torrents simultaneously and only uses one
TCP port. It's fairly low on resource usage and should be perfect for
file distribution sites. Efficient downloads and ease of use makes this
client a good choice for the casual user as well.

WWW: https://github.com/btpd/btpd/wiki

PR:		1
Submitted by:	mat@FreeBSD.org


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@522474 35697150-7ecd-e111-bb59-0022644237b5