Unbreak make makepatch.
The glob in `find ./*` is handled by the shell and actually sorts the
files, so one needs to to ask find to sort the result, using `find -s`.

`find ./* -maxdepth 0` is not equivalent to `find . -maxdepth 0`, the
depth increases by one as `./x` is on level deeper than `.`.

Pointy hat:	bdrewery