Make dependency on linguisttools explicit
The binary found by the cmake configure to check for its availability
found the 'wrapper' installed by misc/qtchooser, which was present
always (in short: it was broken). An upcoming change will change the
behaviour of qtchooser not to install links if the binaries are not
present.

PR:		243443
Reported by:	antoine