Add rubygem-ruby-vips 2.0.17
This gem is a Ruby binding for the libvips image processing library.

Programs that use ruby-vips don't manipulate images directly, instead they
create pipelines of image processing operations building on a source image. When
the end of the pipe is connected to a destination, the whole pipeline executes
at once, streaming the image in parallel from source to destination a section at
a time. Because ruby-vips is parallel, it's quick, and because it doesn't need
to keep entire images in memory, it's light.

WWW: https://github.com/libvips/ruby-vips