math/openblas: actually add libopenblas library to the package on powerpc64 head
On powerpc64 head, libopenblas_ppc970p-r0.3.7.* weren't added to the package, even though the package itself built succesfully. This problem doesn't exist on 12.
Lack of libopenblas_ppc970p-r0.3.7.* made many packages fail.
Add DYNAMIC_ARCH option everywhere since it will cause the post-install target dependent on this option to run and include libopenblas_ppc970p-r0.3.7.* in the package.
This is only to get the necessary libs in the package, the option only works on ppc64le, which we don't support.

Bump PORTREVISION because of package change.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@525231 35697150-7ecd-e111-bb59-0022644237b5