x11-toolkits/p5-Alien-wxWidgets: remove extraneous escapes
post-patch phase of this port uses sed to replace 'wx-config with
'${WX_CONFIG:T}. Previously it escaped the ', but this is not necessary in
this context.

Remove the extraneous escape so that it doesn't get passed through to sed(1)
as an escape of an ordinary character. This is important as the patch in bug
#229925 will make such escapes an error to ease transition into a world
where some escaped-ordinaries will be granted a special meaning.

PR:		240450
Approved by:	koobs (mentor), bapt (mentor)
Approved by:	portmgr (maintainer timeout: 4 months)
MFH:		2020Q1 (future build fix)
Differential Revision:	https://reviews.freebsd.org/D23702