AT&T AST (att/ast) on github has rewound their repo back to ksh93u+. See https://github.com/att/ast/issues/1464 and https://github.com/att/ast/issues/1466.
The reason for the rewind back to 93u+ instead of 93v- was that it was an
abandoned unstable buggy unfinished beta. A full explanation of this can be
found here, https://github.com/att/ast/issues/1466#issuecomment-589942449.

The version number is reverted back to 93u and an EPOCH bump, though not
needed, documents this siesmic shift.