Resurrect, update and adopt misc/xiphos
Resurrect the port and update it from 3.1.6 to a git commit after the
latest release. The commit was selected so that the port does not depend
on python2.7.
Resurrection is possible because the dependency on webkit-gtk3 in the
expired version has been replaced by a dependency on webkit2-gtk3.

PR:		243934
Approved by:	gerald (mentor)